rework DeployApp models
