+++ cloudshell/cp/core/request_actions/get_vm_details.py
@@ -1,11 +1,11 @@
 import json
 from dataclasses import dataclass, field
 
-from cloudshell.cp.core.request_actions.deployed_app import DeployedApp
+from cloudshell.cp.core.request_actions.deployed_vm import DeployedVMRequestActions
 
 
 @dataclass
-class GetVMDetailsRequestActions:
+class GetVMDetailsRequestActions(DeployedVMRequestActions):
     deployed_apps: list = field(default_factory=list)
 
     @classmethod
@@ -18,10 +18,10 @@ class GetVMDetailsRequestActions:
         data = json.loads(request)
         deployed_apps = []
         for item in data["items"]:
-            deployed_app = DeployedApp.from_data(
+            deployed_app_request = cls.from_data(
                 app_request_data=item["appRequestJson"],
                 deployed_app_data=item["deployedAppJson"],
             )
-            deployed_apps.append(deployed_app)
+            deployed_apps.append(deployed_app_request.deployed_app)
 
         return cls(deployed_apps=deployed_apps)
