+++ cloudshell/cp/core/request_actions/models/__init__.py
@@ -3,6 +3,7 @@ from .base import *  # noqa
 from .connectivity import *  # noqa
 from .create_keys import *  # noqa
 from .deploy_app import *  # noqa
+from .deployed_app import *  # noqa
 from .prepare_cloud_infra import *  # noqa
 from .prepare_subnet import *  # noqa
 from .save_restore_app import *  # noqa
